# Small optimization of code