* Examples:
     * CO2-MFI: 9.7 → 8.15 seconds
     * Bae-Mixture: 108.8 → 94.8 seconds
     * NU-2000: 17.20 → 7.5 seconds
     * NPTMC: 223.73 → 175.2 seconds
     * NVT-Gibbs (100 + 100 cycles): 35.3 → 31.45 seconds